progress commit - again, getting closer - dreaded static lifetimes
